Add implied preprocessor definitions
If '_AFXDLL' is in your preprocessor definitions, you will get the
following clang diagnostic error from some windows header(s):

"error: Please use the /MD switch for _AFXDLL builds"

Setting "Use of MFC in a Shared DLL" will get you '_AFXDLL' automatically,
but the code that emits this error is looking for '_DLL'. That doesn't
normally come from preprocessor definitions explicitly, it's controlled by
the project's runtime library setting.

This checks the runtime library setting and adds the appropriate implied
defines to the preprocessor definitions. There are probably a lot more of
these that might turn out to be necessary, but this is the only diagnostic
error I've encountered so far. See the following reference for a list of
candidates:
https://docs.microsoft.com/en-us/cpp/preprocessor/predefined-macros

Fixes #18