Don't use -isystem for non-system include folders (by default)
I found that include-what-you-use (and who knows what else, since IWYU
itself relies on clang's built-in judgement) relies on -isystem to
indicate which include directives are for system includes, so using
-isystem for EVERY include directory results in it spamming angled
brackets all-over the place.

By default, we will use -isystem for platform include directories and -I
for project-specific additional include directories. While the effect
angle brackets and double quotes have on include paths is essentially
implementation defined, this default should mimic the overwhelming
convention and corresponding behavior in MSVC/clang/gcc.

Because there are some projects that mandate exclusive use of angle
brackets for everything, a checkbox is available to enable the old
behavior.